Move topology functions to clkTopology.cpp
