+++ CMakeLists.txt
@@ -35,6 +35,7 @@ add_library(recombo_core STATIC
         src/clkRecombination.cpp
         src/clkValidation.cpp
         src/clkLattice.cpp
+        src/clkTopology.cpp
         src/mmc_config.cpp
         src/mmchain.cpp
         src/newsud.cpp
